Add files via upload
The folder contains scripts required for analyzing the multiplexed data for 24-well acquired via CRITTA software. The scripts have to be executed in the order described in the README for 24-well file. 
An example data output is also provided along with the scripts. 
Please note that the scripts can only operate on Windows OS using Python 2.7